Integration tests splited into submodules so it is easier to test different optional dependencies in isolation
