+++ .github/workflows/integration-tests.yaml
@@ -28,9 +28,10 @@ jobs:
           cache: 'maven'
 
       - name: Compile
-        run: ./mvnw install -Pquick -pl '!integration-test'
+        run: ./mvnw install -Pquick
 
-      - name: Integration Tests (latest - classic - jackson)
-        run: ./mvnw clean verify -P${{ matrix.maven_profiles }} -pl integration-test
+      - name: Integration Tests - Core (latest - classic - jackson)
+        run: ./mvnw clean verify -P it,${{ matrix.maven_profiles }} -pl 'integration-test/core'
 
-
+      - name: Integration Tests - Non-core integrations
+        run: ./mvnw clean verify -P it -pl 'integration-test/metrics,integration-test/zalando'
