+++ integration-test/core/src/main/java/io/quarkiverse/resteasy/problem/GenericExceptionsResource.java
