+++ integration-test/core/src/main/java/io/quarkiverse/resteasy/problem/JaxRsExceptionsResource.java
