+++ integration-test/core/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+# JWT
+mp.jwt.verify.issuer=evry-rfc7807-problem
+mp.jwt.verify.publickey.location=jwt/publicKey.pem
+
+quarkus.default-locale=en-US
