+++ integration-test/core/src/main/resources/jwt/publicKey.pem
