+++ integration-test/core/src/test/java/io/quarkiverse/resteasy/problem/SecurityMappersIT.java
