+++ integration-test/metrics/src/main/java/io/quarkiverse/resteasy/problem/MetricsResource.java
@@ -1,0 +1,26 @@
+package io.quarkiverse.resteasy.problem;
+
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.MediaType;
+
+@Path("/metrics/")
+@Produces(MediaType.APPLICATION_JSON)
+public class MetricsResource {
+
+    @GET
+    @Path("/throw-web-application-exception")
+    public void throwWebApplicationException(@QueryParam("status") int status) {
+        throw new WebApplicationException(new TestRuntimeException(), status);
+    }
+
+    private static final class TestRuntimeException extends RuntimeException {
+        TestRuntimeException() {
+            super("First cause", new RuntimeException("Root cause"));
+        }
+    }
+
+}
