+++ integration-test/metrics/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.resteasy.problem.metrics.enabled=true
+
+quarkus.default-locale=en-US
