+++ integration-test/metrics/src/test/java/io/quarkiverse/resteasy/problem/MetricsNativeIT.java
