--- integration-test/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-quarkus.resteasy.problem.include-mdc-properties=uuid,field-from-configuration,another-field
-quarkus.resteasy.problem.metrics.enabled=true
-
-# JWT
-mp.jwt.verify.issuer=evry-rfc7807-problem
-mp.jwt.verify.publickey.location=jwt/publicKey.pem
-
-quarkus.default-locale=en-US
