+++ integration-test/zalando/src/test/java/io/quarkiverse/resteasy/problem/ZalandoProblemMapperIT.java
