+++ run-jvm-tests
@@ -15,9 +15,10 @@ fi
 set -x
 set -e
 
-./mvnw clean install -Pquick -pl '!integration-test'
+./mvnw clean install -Pquick
 
-./mvnw clean verify -Pjackson-classic -pl integration-test
-./mvnw clean verify -Pjsonb-classic -pl integration-test
-./mvnw clean verify -Pjackson-reactive -pl integration-test
-./mvnw clean verify -Pjsonb-reactive -pl integration-test
+./mvnw clean verify -P it,jackson-classic -pl 'integration-test/core'
+./mvnw clean verify -P it,jsonb-classic -pl 'integration-test/core'
+./mvnw clean verify -P it,jackson-reactive -pl 'integration-test/core'
+./mvnw clean verify -P it,jsonb-reactive -pl 'integration-test/core'
+./mvnw clean verify -P it -pl 'integration-test/metrics,integration-test/zalando'
