+++ runtime/src/main/java/io/quarkiverse/resteasy/problem/HttpProblem.java
@@ -13,7 +13,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import jakarta.annotation.Nullable;
-
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
